evalengine/integration: move all tests into their own package
