+++ go/vt/vtgate/evalengine/integration/fuzz_test.go
@@ -90,6 +90,7 @@ var (
 		regexp.MustCompile(`You have an error in your SQL syntax; (.*?)`),
 		regexp.MustCompile(`Cannot convert string '(.*?)' from \w+ to \w+`),
 		regexp.MustCompile(`Invalid JSON text in argument (\d+) to function (\w+): (.*?)`),
+		regexp.MustCompile(`Illegal mix of collations`),
 	}
 )
 
@@ -119,7 +120,7 @@ func errorsMatch(remote, local error) bool {
 	return false
 }
 
-func safeEvaluate(env *evalengine.ExpressionEnv, query string) (evalengine.EvalResult, sqltypes.Type, error) {
+func evaluateLocalEvalengine(env *evalengine.ExpressionEnv, query string) (evalengine.EvalResult, sqltypes.Type, error) {
 	stmt, err := sqlparser.Parse(query)
 	if err != nil {
 		return evalengine.EvalResult{}, 0, err
@@ -137,23 +138,22 @@ func safeEvaluate(env *evalengine.ExpressionEnv, query string) (evalengine.EvalR
 		return
 	}()
 
-	var eval evalengine.EvalResult
-	var tt sqltypes.Type
-	if err == nil {
-		eval, err = func() (eval evalengine.EvalResult, err error) {
-			defer func() {
-				if r := recover(); r != nil {
-					err = fmt.Errorf("PANIC: %v", r)
-				}
-			}()
-			eval, err = env.Evaluate(local)
-			if err == nil && debugCheckTypes {
-				tt, err = env.TypeOf(local)
-			}
-			return
-		}()
+	if err != nil {
+		return evalengine.EvalResult{}, 0, err
 	}
-	return eval, tt, err
+
+	return func() (eval evalengine.EvalResult, tt sqltypes.Type, err error) {
+		defer func() {
+			if r := recover(); r != nil {
+				err = fmt.Errorf("PANIC: %v", r)
+			}
+		}()
+		eval, err = env.Evaluate(local)
+		if err == nil && debugCheckTypes {
+			tt, err = env.TypeOf(local)
+		}
+		return
+	}()
 }
 
 const syntaxErr = `You have an error in your SQL syntax; (errno 1064) (sqlstate 42000) during query: SQL`
@@ -183,7 +183,7 @@ func TestGenerateFuzzCases(t *testing.T) {
 		query := "SELECT " + sqlparser.String(expr)
 
 		env := evalengine.EnvWithBindVars(nil, 255)
-		eval, _, localErr := safeEvaluate(env, query)
+		eval, _, localErr := evaluateLocalEvalengine(env, query)
 		remote, remoteErr := conn.ExecuteFetch(query, 1, false)
 
 		if localErr != nil && strings.Contains(localErr.Error(), "syntax error at position") {
